Update Creating_Dataset_2
Run each section individually. Before running "Apply pre-processing techniques" just check the video you recorded looks good with no lighting issues etc and then continue.

When using the file make sure to:
Change the video name and the file location (line 16, 33, 34) to the corresponding hand gesture. In the script I have it as Descend.
Ensure the file location of the images (line 49) is in a subfolder of the hand gesture folder (where the videos are) just for ease of use for when Liam puts the images into the Deep Learning network.

Finally, remember if you have more than 1 webcam change line 4 to the right webcam.